+++ COLA.hh
@@ -21,7 +21,7 @@
 #ifndef COLA_COLA_HH
 #define COLA_COLA_HH
 
-#include "LorentzVector.hh"
+#include "COLA/EventData.hh"
 
 #include <memory>
 #include <optional>
@@ -34,108 +34,7 @@ namespace tinyxml2 {
 }  // namespace tinyxml2
 
 namespace cola {
-  using LorentzVector = LorentzVectorImpl<double>;
-
-  /** A typedef representing mass and charge of a nucleon.
-   */
-  using AZ = std::pair<uint16_t, uint16_t>;
-
-  /** PDG code to AZ converter.
-   *  **WARNING:** this function is intended to process heavy ions PDG codes.
-   *  @param pdgCode PDG code of the ion.
-   *  @return AZ of the ion.
-   */
-  AZ PdgToAZ(int pdgCode);
-
-  /** AZ to PDG code convverter.
-   *  @param data AZ of the ion.
-   *  @return PDG code of the ion
-   */
-  int AZToPdg(AZ data);
-
-  /** \defgroup Data Data Classes and supporting methods.
-   * @{
-   */
-
-  /** Particle class by generator output.
-   *  This enum represents various outcomes of a generator event for every particle.
-   */
-  enum class ParticleClass : char {
-    kProduced,    /**< A particle that was not present in the starting nuclei. */
-    kElasticA,    /**< A particle that was present in the projectile nucleus and has experienced only elastic
-                   * interactions.
-                   */
-    kElasticB,    /**< A particle that was present in the target nucleus and has experienced only elastic interactions.
-                   */
-    kNonelasticA, /**< A particle that was present in the projectile nucleus and has experienced at least one
-                    non-elastic interaction. */
-    kNonelasticB, /**< A particle that was present in the target nucleus and has experienced at least one
-                    non-elastic interaction. */
-    kSpectatorA,  /**< A particle that was present in the projectile nucleus and hasn't experienced any interactions.
-                   */
-    kSpectatorB   /**< A particle that was present in the projectile nucleus and hasn't experienced any interactions.
-                   */
-  };
-
-  /** Particle data.
-   *  A structure representing data about a single particle
-   */
-  struct Particle {
-    AZ GetAZ() const;
-
-    LorentzVector position; /**< Position <t, x, y, z> vector. */
-
-    LorentzVector momentum; /**< Momentum <e, x, y, z> vector. */
-
-    int pdgCode;          /**< PDG code of the particle. */
-    ParticleClass pClass; /**< Data about particle origin. See ParticleClass for more info.*/
-  };
-
-  /**
-   * Convenient typedef for Particle vector.
-   */
-  using EventParticles = std::vector<Particle>;
-
-  /** Initial state data.
-   *  This structure contains data about initial state of any given event.
-   */
-  struct EventIniState {
-    int pdgCodeA; /**< PDF code of the projectile. */
-    int pdgCodeB; /**< PDF code of the target. */
-
-    double pZA;    /** Axial momentum of the projectile */
-    double pZB;    /** Axial momentum of the target */
-    double energy; /** Incidental energy of the event. Depending on pZB being zero, this is either \f$E/A\f$ of
-                      target or \f$\sqrt{s_{NN}}\f$. */
-
-    float sectNN; /** Nucleon-Nucleon cross section from generator. */
-    float b;      /** Impact parameter of the event. */
-
-    int nColl;   /** Diagnostic. Total number of collisions. */
-    int nCollPP; /** Diagnostic. Number of proton-proton. */
-    int nCollPN; /** Diagnostic. Number of proton-neutron collisions. */
-    int nCollNN; /** Diagnostic. Number of neutron-neutron collisions. */
-    int nPart;   /** Diagnostic. Total number of participants. */
-    int nPartA;  /** Diagnostic. Number of participants from the projectile nucleus. */
-    int nPartB;  /** Diagnostic. Number of participants from the target nucleus. */
-
-    float phiRotA;   /** Diagnostic. Polar angle \f$\phi\f$ of rotation of the projectile nucleon. */
-    float thetaRotA; /** Diagnostic. Polar angle \f&\Theta\f$ of rotation of the projectile nucleon. */
-    float phiRotB;   /** Diagnostic. Polar angle \f$\phi\f$ of rotation of the target nucleon. */
-    float thetaRotB; /** Diagnostic. Polar angle \f$\Theta\f$ of rotation of the target nucleon. */
-
-    EventParticles iniStateParticles; /** The array of all Particles just before the event. */
-  };
-
-  /** A structure combining EventIniState and EventParticles of the event.
-   */
-  struct EventData {
-    EventIniState iniState;
-    EventParticles particles;
-  };
-
-  /** @}
-   * \defgroup Interface Pure abstract classes used for dependency injection.
+  /** \defgroup Interface Pure abstract classes used for dependency injection.
    * @{
    */
 
@@ -271,13 +170,17 @@ namespace cola {
     FilterType GetFilterType() const final { return FilterType::kWriter; }
   };
 
-  template <typename Filter> class GenericFactory : public VFactory {
+  template <typename Filter>
+  class GenericFactory : public VFactory {
    private:
-    template <typename, typename = void> struct HasName : std::false_type {};
+    template <typename, typename = void>
+    struct HasName : std::false_type {};
 
-    template <typename T> struct HasName<T, std::void_t<decltype(T::kName)>> : std::true_type {};
+    template <typename T>
+    struct HasName<T, std::void_t<decltype(T::kName)>> : std::true_type {};
 
-    template <typename T> static constexpr bool kHasName = HasName<T>::value;
+    template <typename T>
+    static constexpr bool kHasName = HasName<T>::value;
 
    public:
     GenericFactory() {
@@ -419,9 +322,11 @@ namespace cola {
     virtual FactoryMap DoGetModuleFilters() const = 0;
   };
 
-  template <typename... FilterTypes> class GenericModule : public VModule {
+  template <typename... FilterTypes>
+  class GenericModule : public VModule {
    private:
-    template <typename HeadType, typename... Types> static void AddFilter(FactoryMap& factories) {
+    template <typename HeadType, typename... Types>
+    static void AddFilter(FactoryMap& factories) {
       static_assert(std::is_base_of_v<VFactory, HeadType>, "all types must be Factories");
 
       {
